--- src/main/kotlin/com/knu/mockin/model/dto/kisrequest/TestRequestParameter.kt
@@ -1,8 +1,0 @@
-package com.knu.mockin.model.dto.kisrequest
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class TestRequestParameter (
-        @JsonProperty("FID_COND_MRKT_DIV_CODE") val fidCondMrktDivCode : String,
-        @JsonProperty("FID_INPUT_ISCD") val fidInputIscd : String,
-)
