--- src/main/kotlin/com/knu/mockin/model/dto/kisresponse/TestResponseDto.kt
@@ -1,8 +1,0 @@
-package com.knu.mockin.model.dto.kisresponse
-
-import com.knu.mockin.model.dto.kisheader.response.KISOverSeaResponseHeaderDto
-
-data class TestResponseDto (
-        val header: KISOverSeaResponseHeaderDto, // 헤더 (필수)
-        val body: TestResponseBodyDto // 본문 (필수)
-)
