+++ src/main/kotlin/com/knu/mockin/service/CurrentPriceService.kt
@@ -1,21 +1,22 @@
 package com.knu.mockin.service
 
-import com.knu.mockin.kisclient.KISOverSeaClient
+import com.knu.mockin.kisclient.KISTradingClient
 import com.knu.mockin.model.dto.kisheader.request.KISOverSeaRequestHeaderDto
 import com.knu.mockin.model.dto.kisrequest.basicprice.currentprice.KISCurrentPriceRequestParameter
-import com.knu.mockin.model.dto.kisresponse.basicprice.currentprice.KISCurrentPriceResponseDto
+import com.knu.mockin.model.dto.kisresponse.basicprice.currentprice.KISCurrentPriceResponseBodyDto
 import com.knu.mockin.model.enum.TradeId
 import com.knu.mockin.repository.UserRepository
+import kotlinx.coroutines.reactive.awaitFirst
+import kotlinx.coroutines.reactive.awaitSingle
 import org.springframework.stereotype.Service
-import reactor.core.publisher.Mono
 
 @Service
 class CurrentPriceService(
-        private val kisOverSeaClient: KISOverSeaClient,
+        private val kisTradingClient: KISTradingClient,
         private val userRepository: UserRepository
 ) {
-    fun getCurrentPrice(): Mono<KISCurrentPriceResponseDto> {
-        return userRepository.findById(1).flatMap { user ->
+    suspend fun getCurrentPrice(): KISCurrentPriceResponseBodyDto {
+            val user = userRepository.findById(1).awaitFirst();
             // Request header 생성
             val kisOverSeaRequestHeaderDto = KISOverSeaRequestHeaderDto(
                     authorization = "Bearer ${user.token}",
@@ -26,13 +27,12 @@ class CurrentPriceService(
 
             // 요청 파라미터 생성
             val requestParameter = KISCurrentPriceRequestParameter(
-                    AUTH = "",  // 필요한 경우 AUTH 값을 설정
+                      // 필요한 경우 AUTH 값을 설정
                     EXCD = "NAS",
                     SYMB = "TSLA"
             )
 
             // KIS API 호출
-            kisOverSeaClient.getCurrentPrice(kisOverSeaRequestHeaderDto, requestParameter)
-        }
+            return kisTradingClient.getCurrentPrice(kisOverSeaRequestHeaderDto, requestParameter).awaitSingle()
     }
 }
